+++ common/extension/tps_limit.go
@@ -18,20 +18,20 @@
 package extension
 
 import (
-	"github.com/apache/dubbo-go/filter/impl/tps/intf"
+	"github.com/apache/dubbo-go/filter/impl/tps"
 )
 
 var (
-	tpsLimitStrategy            = make(map[string]func(rate int, interval int) intf.TpsLimitStrategy)
-	tpsLimiter                  = make(map[string]func() intf.TpsLimiter)
-	tpsRejectedExecutionHandler = make(map[string]func() intf.RejectedExecutionHandler)
+	tpsLimitStrategy            = make(map[string]func(rate int, interval int) tps.TpsLimitStrategy)
+	tpsLimiter                  = make(map[string]func() tps.TpsLimiter)
+	tpsRejectedExecutionHandler = make(map[string]func() tps.RejectedExecutionHandler)
 )
 
-func SetTpsLimiter(name string, creator func() intf.TpsLimiter) {
+func SetTpsLimiter(name string, creator func() tps.TpsLimiter) {
 	tpsLimiter[name] = creator
 }
 
-func GetTpsLimiter(name string) intf.TpsLimiter {
+func GetTpsLimiter(name string) tps.TpsLimiter {
 	creator, ok := tpsLimiter[name]
 	if !ok {
 		panic("TpsLimiter for " + name + " is not existing, make sure you have import the package " +
@@ -40,11 +40,11 @@ func GetTpsLimiter(name string) intf.TpsLimiter {
 	return creator()
 }
 
-func SetTpsLimitStrategy(name string, creator func(rate int, interval int) intf.TpsLimitStrategy) {
+func SetTpsLimitStrategy(name string, creator func(rate int, interval int) tps.TpsLimitStrategy) {
 	tpsLimitStrategy[name] = creator
 }
 
-func GetTpsLimitStrategyCreator(name string) func(rate int, interval int) intf.TpsLimitStrategy {
+func GetTpsLimitStrategyCreator(name string) func(rate int, interval int) tps.TpsLimitStrategy {
 	creator, ok := tpsLimitStrategy[name]
 	if !ok {
 		panic("TpsLimitStrategy for " + name + " is not existing, make sure you have import the package " +
@@ -53,11 +53,11 @@ func GetTpsLimitStrategyCreator(name string) func(rate int, interval int) intf.T
 	return creator
 }
 
-func SetTpsRejectedExecutionHandler(name string, creator func() intf.RejectedExecutionHandler) {
+func SetTpsRejectedExecutionHandler(name string, creator func() tps.RejectedExecutionHandler) {
 	tpsRejectedExecutionHandler[name] = creator
 }
 
-func GetTpsRejectedExecutionHandler(name string) intf.RejectedExecutionHandler {
+func GetTpsRejectedExecutionHandler(name string) tps.RejectedExecutionHandler {
 	creator, ok := tpsRejectedExecutionHandler[name]
 	if !ok {
 		panic("TpsRejectedExecutionHandler for " + name + " is not existing, make sure you have import the package " +
