+++ filter/impl/tps/impl/tps_limit_strategy_mock.go
@@ -18,7 +18,7 @@
 // Source: tps_limit_strategy.go
 
 // Package filter is a generated GoMock package.
-package tps
+package impl
 
 import (
 	gomock "github.com/golang/mock/gomock"
