+++ filter/impl/tps/impl/tps_limit_thread_safe_fix_window_strategy_test.go
@@ -15,22 +15,29 @@
  * limitations under the License.
  */
 
-package intf
+package impl
 
-/*
- * please register your implementation by invoking SetTpsLimitStrategy
- * "UserProvider":
- *   registry: "hangzhouzk"
- *   protocol : "dubbo"
- *   interface : "com.ikurento.user.UserProvider"
- *   ... # other configuration
- *   tps.limiter: "method-service" # the name of limiter
- *   tps.limit.strategy: "name of implementation" # service-level
- *   methods:
- *    - name: "GetUser"
- *      tps.interval: 3000
- *      tps.limit.strategy: "name of implementation" # method-level
- */
-type TpsLimitStrategy interface {
-	IsAllowable() bool
+import (
+	"testing"
+	"time"
+)
+
+import (
+	"github.com/stretchr/testify/assert"
+)
+
+func TestThreadSafeFixedWindowTpsLimitStrategyImpl_IsAllowable(t *testing.T) {
+	strategy := NewThreadSafeFixedWindowTpsLimitStrategyImpl(2, 60000)
+	assert.True(t, strategy.IsAllowable())
+	assert.True(t, strategy.IsAllowable())
+	assert.False(t, strategy.IsAllowable())
+
+	strategy = NewThreadSafeFixedWindowTpsLimitStrategyImpl(2, 2000)
+	assert.True(t, strategy.IsAllowable())
+	assert.True(t, strategy.IsAllowable())
+	assert.False(t, strategy.IsAllowable())
+	time.Sleep(time.Duration(2100 * 1000))
+	assert.True(t, strategy.IsAllowable())
+	assert.True(t, strategy.IsAllowable())
+	assert.False(t, strategy.IsAllowable())
 }
