+++ filter/impl/tps/tps_limit_strategy.go
@@ -17,19 +17,20 @@
 
 package tps
 
-import (
-	"net/url"
-	"testing"
-)
-import (
-	"github.com/apache/dubbo-go/common"
-	"github.com/apache/dubbo-go/common/constant"
-)
-
-func TestOnlyLogRejectedExecutionHandler_RejectedExecution(t *testing.T) {
-	handler := GetOnlyLogRejectedExecutionHandler()
-	invokeUrl := common.NewURLWithOptions(
-		common.WithParams(url.Values{}),
-		common.WithParamsValue(constant.INTERFACE_KEY, "methodName"))
-	handler.RejectedExecution(*invokeUrl, nil)
+/*
+ * please register your implementation by invoking SetTpsLimitStrategy
+ * "UserProvider":
+ *   registry: "hangzhouzk"
+ *   protocol : "dubbo"
+ *   interface : "com.ikurento.user.UserProvider"
+ *   ... # other configuration
+ *   tps.limiter: "method-service" # the name of limiter
+ *   tps.limit.strategy: "name of implementation" # service-level
+ *   methods:
+ *    - name: "GetUser"
+ *      tps.interval: 3000
+ *      tps.limit.strategy: "name of implementation" # method-level
+ */
+type TpsLimitStrategy interface {
+	IsAllowable() bool
 }
